+++ src/main/java/com/github/jonpeterson/jackson/module/versioning/VersioningBeanDeserializationModifier.java
@@ -29,13 +29,11 @@ import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.deser.BeanDeserializerModifier;
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 import com.fasterxml.jackson.databind.introspect.BeanPropertyDefinition;
-import static com.github.jonpeterson.jackson.module.versioning.VersioningBeanSerializationModifier.getSerializeToVersionProperty;
 
 class VersioningBeanDeserializationModifier extends BeanDeserializerModifier {
 
     // here just to make generics work without warnings
-    private static <T> VersionedModelDeserializer<T> createVersioningDeserializer(StdDeserializer<T> deserializer, JsonVersionedModel jsonVersionedModel, 
-            BeanPropertyDefinition serializeToVersionProperty) {
+    private static <T> VersionedModelDeserializer<T> createVersioningDeserializer(StdDeserializer<T> deserializer, JsonVersionedModel jsonVersionedModel, BeanPropertyDefinition serializeToVersionProperty) {
         return new VersionedModelDeserializer<T>(deserializer, jsonVersionedModel, serializeToVersionProperty);
     }
 
@@ -45,8 +43,11 @@ class VersioningBeanDeserializationModifier extends BeanDeserializerModifier {
         if(deserializer instanceof StdDeserializer) {
             JsonVersionedModel jsonVersionedModel = beanDescription.getClassAnnotations().get(JsonVersionedModel.class);
             if(jsonVersionedModel != null)
-                return createVersioningDeserializer((StdDeserializer)deserializer, jsonVersionedModel,
-                        getSerializeToVersionProperty(beanDescription));
+                return createVersioningDeserializer(
+                    (StdDeserializer)deserializer,
+                    jsonVersionedModel,
+                    VersionedModelUtils.getSerializeToVersionProperty(beanDescription)
+                );
         }
 
         return deserializer;